Cleaned up the code for the checkin process
